Move jukebox into its own window
